@@ -3,6 +3,7 @@
   "version": "3.9.2",
   "description": "My ADHD Website",
   "homepage": "https://myadhd.com.au",
+  "type": "module",
   "bugs": {
     "url": "https://github.com/renansigolo/my-adhd-website/issues",
     "email": "renan.sigolo@gmail.com"
@@ -42,12 +43,12 @@
     "@tailwindcss/forms": "0.5.10",
     "@vercel/analytics": "1.5.0",
     "@vercel/speed-insights": "1.2.0",
-    "firebase": "11.9.1",
+    "firebase": "11.10.0",
     "next": "15.4.0",
     "react": "19.1.0",
     "react-dom": "19.1.0",
     "react-firebase-hooks": "5.1.1",
-    "react-hook-form": "7.58.0",
+    "react-hook-form": "7.60.0",
     "react-hot-toast": "2.5.2",
     "react-intl": "7.1.11",
     "sharp": "0.34.2"
@@ -55,27 +56,24 @@
   "devDependencies": {
     "@commitlint/cli": "19.8.1",
     "@commitlint/config-conventional": "19.8.1",
+    "@eslint/eslintrc": "3.3.1",
     "@semantic-release/changelog": "6.0.3",
     "@semantic-release/git": "10.0.1",
-    "@tailwindcss/postcss": "4.1.10",
+    "@tailwindcss/postcss": "4.1.11",
     "@types/gtag.js": "0.0.20",
     "chance": "1.1.13",
-    "concurrently": "9.1.2",
-    "cypress": "14.4.1",
-    "dotenv": "16.5.0",
-    "eslint": "8.57.1",
-    "eslint-config-next": "15.3.3",
+    "concurrently": "9.2.0",
+    "cypress": "14.5.1",
+    "dotenv": "17.1.0",
+    "eslint": "9.30.1",
+    "eslint-config-next": "15.3.5",
     "eslint-config-prettier": "10.1.5",
     "husky": "9.1.7",
     "next-sitemap": "4.2.3",
     "postcss": "8.5.6",
-    "prettier": "3.5.3",
-    "prettier-plugin-tailwindcss": "0.6.12",
-    "semantic-release": "24.2.5",
-    "tailwindcss": "4.1.10"
-  },
-  "engines": {
-    "node": "^20.11.1 || ^22.12.0",
-    "npm": ">=10.2.4"
+    "prettier": "3.6.2",
+    "prettier-plugin-tailwindcss": "0.6.13",
+    "semantic-release": "24.2.6",
+    "tailwindcss": "4.1.11"
   }
 }
