@@ -1,5 +1,5 @@
-/** @type {import('semantic-release').Options} */
-module.exports = {
+/** @type {import('semantic-release').GlobalConfig} */
+export default {
   branches: ["main"],
   plugins: [
     "@semantic-release/commit-analyzer",
