@@ -1,5 +1,6 @@
 /** @type {import('next-sitemap').IConfig} */
-module.exports = {
+const config = {
   siteUrl: process.env.SITE_URL || "https://myadhd.app",
   generateRobotsTxt: true,
 }
+export default config
