@@ -1,5 +1,5 @@
 /** @type {import('next').NextConfig} */
-module.exports = {
+const nextConfig = {
   reactStrictMode: true,
   images: {
     remotePatterns: [
@@ -14,3 +14,5 @@ module.exports = {
     defaultLocale: "en",
   },
 }
+
+export default nextConfig
