@@ -5,7 +5,7 @@ export function Disclaimer() {
   return (
     <div className="bg-linear-to-tl from-purple-800 to-pink-500 text-center">
       <ContainerSection>
-        <p className="text-2xl text-white">
+        <p className="text-2xl text-white" aria-required>
           <FormattedMessage
             defaultMessage="This app is for educational purposes only and does not reflect a medical opinion. ADHD can only be diagnosed by a medical professional!"
             id="disclaimer.title"
